requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
swanlab

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils~=1.0.7
pre-commit      # hooks for applying linters on commit
rich~=13.9.4
pytest~=8.4.2
# sh            # for running bash commands in some tests (linux/macos only)

pandas~=2.3.3
numpy~=2.0.2
scikit-learn~=1.6.1
omegaconf~=2.3.0
joblib~=1.5.2
seaborn~=0.13.2
packaging~=25.0
setuptools~=68.2.0